Create Problem_48.cpp
You are packing for a vacation on the sea side and you are going to carry only one bag with 
capacity C (1 <= C <= 1000). You also have N (1<= N <= 1000) items that you might want to take 
with you to the sea side. Unfortunately you can not fit all of them in the knapsack so you will 
have to choose. For each item you are given its size and its value. You want to maximize the total 
value of all the items you are going to bring. What is this maximum total value?
Example :
	Input : N = 5 C = 4
			P = {8, 4, 0, 5, 3}
			W = {1, 2, 3, 2, 2}
			
	Output: 13